Put ACME request handler to its own file
